split out emitUtils from realtimeUtils
